refactor(frontend): extract BulkActionBar + Pagination + row badges from contacts page (778 -> 435 lines)
